move doctest_unittest configurbale and moved to test_tools
